+++ fenbux/univariate/continuous/_pareto.py
@@ -180,9 +180,7 @@ def _sf(d: Pareto, x):
 
 
 @_rand_impl.dispatch
-def _rand(
-    d: Pareto, key: KeyArray, shape: Shape = (), dtype: DTypeLikeFloat = jnp.float_
-):
+def _rand(d: Pareto, key: KeyArray, shape: Shape = (), dtype: DTypeLikeFloat = float):
     d = d.broadcast_params()
     _key_tree = split_tree(key, d.shape)
     return jtu.tree_map(
