+++ fenbux/univariate/continuous/_weibull.py
@@ -196,7 +196,7 @@ def _sf(d: Weibull, x: PyTreeVar):
 
 @_rand_impl.dispatch
 def _rand(
-    d: Weibull, key: KeyArray, shape: Shape = (), dtype: DTypeLikeFloat = jnp.float_
+    d: Weibull, key: KeyArray, shape: Shape = (), dtype: DTypeLikeFloat = float
 ):
     d = d.broadcast_params()
     _key_tree = split_tree(key, d.shape)
